Fix GenerateMaze & ConfigMaze class
In the previous codebase, ConfigMaze was expected to take in input from user and generate the maze and then GenerateMaze has a constructor that takes in the getters from ConfigMaze to generate a Symbol Maze. This logic was flawed because it violated SRP but had many code smells. When trying to test the functionality, GenerateMaze required the maze dimensions - absolutely flawed logic.

Refactored to allow for better logic between classes; GenerateMaze provides the maze + dimensions and ConfigMaze simply takes input from the user and parses the Maze. Now follows SRP and reduced codesmell.